refactor: common-build functions
