+++ mermin/Cargo.toml
@@ -30,7 +30,7 @@ clap = { workspace = true, features = ["derive"] }
 [build-dependencies]
 anyhow = { workspace = true }
 aya-build = { workspace = true }
-toml = { workspace = true }
+common-build = { path = "../common-build" }
 # TODO(https://github.com/rust-lang/cargo/issues/12375): this should be an artifact dependency, but
 # it's not possible to tell cargo to use `-Z build-std` to build it. We cargo-in-cargo in the build
 # script to build this, but we want to teach cargo about the dependecy so that cache invalidation
